fix: address review comments on JUnit 5 migration PR
- Update sbt-jupiter-interface from 0.11.0 to 0.18.0 with new org
  (net.aichler -> com.github.sbt.junit)
- Remove redundant testFrameworks line (plugin handles this)
- Add JUnitJupiterRouteTest and deprecate JUnit5RouteTest
- Update ActorSystemExtension doc to reference new class name
